Releaser added; proc improvement

+++ src/testing/procs.go
@@ -2,21 +2,39 @@ package tltest
 
 import (
 	"fmt"
+	"log"
 	"os"
 	"os/exec"
 	"time"
 )
 
+type procServer struct {
+	id     int
+	phy    string
+	dbpath string
+	cmd    *exec.Cmd
+}
+
 func procStartCluster(numServers int) []testServer {
+	dbTestFolder := ""
+	if exists("/mnt/dbs/") {
+		dbTestFolder = "/mnt/dbs/"
+	}
+
 	l := make([]testServer, numServers)
-	l[0] = new(procServer)
+	ps := new(procServer)
+	ps.phy = string("127.0.0.1" + ":" + fmt.Sprint(10000))
+	ps.dbpath = dbTestFolder + "testDB" + fmt.Sprint(0)
+	l[0] = ps
 	for i := 1; i < numServers; i++ {
-		l[i] = new(procServer)
+		ps = new(procServer)
+		ps.id = i
+		ps.phy = string("127.0.0.1" + ":" + fmt.Sprint(10000+i))
+		ps.dbpath = dbTestFolder + "testDB" + fmt.Sprint(i)
+		l[i] = ps
 	}
 	return l
 }
-
-var procID = 0
 
 func exists(path string) bool {
 	_, err := os.Stat(path)
@@ -29,32 +47,20 @@ func exists(path string) bool {
 	return true
 }
 
-type procServer struct {
-	phy    string
-	dbpath string
-	cmd    *exec.Cmd
-}
-
 func (ps *procServer) addr() string {
 	return ps.phy
 }
 
 func (ps *procServer) create(numChunks, redundancy int) string {
+	ps.kill()
 	exec.Command("killall", "-q", "treeless").Run()
-	dbTestFolder := ""
-	if exists("/mnt/dbs/") {
-		dbTestFolder = "/mnt/dbs/"
-	}
-	ps.dbpath = dbTestFolder + "testDB" + fmt.Sprint(procID)
-	ps.dbpath = "testDB" + fmt.Sprint(procID)
 	ps.cmd = exec.Command("./treeless", "-create", "-port",
-		fmt.Sprint(10000+procID), "-dbpath", ps.dbpath, "-localip", "127.0.0.1",
+		fmt.Sprint(10000+ps.id), "-dbpath", ps.dbpath, "-localip", "127.0.0.1",
 		"-redundancy", fmt.Sprint(redundancy), "-chunks", fmt.Sprint(numChunks))
 	if true {
 		ps.cmd.Stdout = os.Stdout
 		ps.cmd.Stderr = os.Stderr
 	}
-	procID++
 	err := ps.cmd.Start()
 	cmdCopy := ps.cmd
 	go func() {
@@ -63,25 +69,17 @@ func (ps *procServer) create(numChunks, redundancy int) string {
 	if err != nil {
 		panic(err)
 	}
-	ps.phy = string("127.0.0.1" + ":" + fmt.Sprint(10000+procID-1))
 	waitForServer(ps.phy)
 	return ps.phy
 }
 
 func (ps *procServer) assoc(addr string) string {
-	dbTestFolder := ""
-	if exists("/mnt/dbs/") {
-		dbTestFolder = "/mnt/dbs/"
-	}
-	ps.dbpath = dbTestFolder + "testDB" + fmt.Sprint(procID)
-	ps.dbpath = "testDB" + fmt.Sprint(procID)
 	ps.cmd = exec.Command("./treeless", "-assoc", addr, "-port",
-		fmt.Sprint(10000+procID), "-dbpath", ps.dbpath, "-localip", "127.0.0.1") //, "-cpuprofile"
+		fmt.Sprint(10000+ps.id), "-dbpath", ps.dbpath, "-localip", "127.0.0.1") //, "-cpuprofile"
 	if true {
 		ps.cmd.Stdout = os.Stdout
 		ps.cmd.Stderr = os.Stderr
 	}
-	procID++
 	err := ps.cmd.Start()
 	cmdCopy := ps.cmd
 	go func() {
@@ -90,15 +88,17 @@ func (ps *procServer) assoc(addr string) string {
 	if err != nil {
 		panic(err)
 	}
-	ps.phy = string("127.0.0.1" + ":" + fmt.Sprint(10000+procID-1))
 	waitForServer(ps.phy)
 	return ps.phy
 }
 
 func (ps *procServer) kill() {
-	ps.cmd.Process.Signal(os.Interrupt)
-	time.Sleep(time.Millisecond * 10)
-	os.RemoveAll(ps.dbpath)
+	if ps.cmd != nil {
+		ps.cmd.Process.Signal(os.Kill)
+		log.Println("Killed", ps.dbpath)
+		time.Sleep(time.Millisecond * 10)
+		os.RemoveAll(ps.dbpath)
+	}
 }
 
 func (ps *procServer) restart() {
